refactor(phase1): extract generators from bin/cli.js
Part of #89 - Phase 1: Extract generators

Extracted 6 generator functions to lib/generators/:
- writeConfig + enrichConfigWithDomains → config-json.js
- writeGuideMd → guide-md.js
- writeAgentsMd → agents-md.js
- writeCursorRules → cursorrules.js
- writeEslintConfig → eslint-config.js

Progress:
- bin/cli.js: 769 → 537 lines (-232 lines, -30%)
- All 554 tests passing

This is the first of 6 PRs to refactor bin/cli.js to a thin router (~40 lines)
