+++ lib/generators/eslint-config.js
@@ -1,0 +1,42 @@
+'use strict';
+
+const fs = require('fs');
+const path = require('path');
+
+function writeEslintConfig(cwd, cfg) {
+  const file = path.join(cwd, 'eslint.config.js');
+  if (fs.existsSync(file) && !process.env.FORCE_ESLINT_CONFIG) {
+    console.log(`Found existing ${file} — set FORCE_ESLINT_CONFIG=1 to overwrite.`);
+    return;
+  }
+  
+  // Check if architecture guardrails are enabled
+  let archRulesConfig = '';
+  const hasArchGuardrails = !!(cfg && cfg.architecture);
+  if (hasArchGuardrails) {
+    try {
+      const { generateArchitectureRules } = require(path.join(__dirname, 'eslint-arch-config'));
+      const { rules } = generateArchitectureRules(cfg.architecture);
+      
+      // Convert rules to ESLint config format
+      const rulesStr = Object.entries(rules).map(([rule, config]) => {
+        return `      '${rule}': ${JSON.stringify(config)},`;
+      }).join('\n');
+      
+      archRulesConfig = `\n      // Architecture guardrails\n${rulesStr}`;
+    } catch (err) {
+      console.warn('Warning: Failed to generate architecture rules:', err.message);
+    }
+  }
+  
+  // When architecture guardrails are enabled, skip overlapping AI-friendly rules
+  const aiFriendlyRules = hasArchGuardrails
+    ? ''
+    : `      // AI-friendly\n      'complexity': ['warn', 15],\n      'max-depth': ['warn', 4],\n      'max-lines-per-function': ['warn', 100],\n`;
+  
+  const content = `// Generated by eslint-plugin-ai-code-snifftest init\n// NOTE: Add "type": "module" to package.json to avoid ESLint module warnings\nimport js from '@eslint/js';\nimport globals from 'globals';\nimport aiSnifftest from 'eslint-plugin-ai-code-snifftest';\n\nexport default [\n  js.configs.recommended,\n  {\n    files: ['**/*.js'],\n    languageOptions: {\n      globals: {\n        ...globals.node\n      }\n    },\n    plugins: { 'ai-code-snifftest': aiSnifftest },\n    rules: {\n      // Baseline\n      'no-unused-vars': ['warn', { argsIgnorePattern: '^_' }],\n      'no-undef': 'error',\n      'prefer-const': 'warn',\n      'no-var': 'error',\n      // Consistency\n      'quotes': ['warn', 'single', { avoidEscape: true }],\n      'semi': ['warn', 'always'],\n      'eqeqeq': ['error', 'always'],\n${aiFriendlyRules}      // Naming (basic)\n      'camelcase': ['error', { properties: 'always' }],\n      // Domain-specific\n      'ai-code-snifftest/no-redundant-calculations': 'warn',\n      'ai-code-snifftest/no-equivalent-branches': 'warn',\n      'ai-code-snifftest/prefer-simpler-logic': 'warn',\n      'ai-code-snifftest/no-redundant-conditionals': 'warn',\n      'ai-code-snifftest/no-unnecessary-abstraction': 'warn',\n      'ai-code-snifftest/no-generic-names': 'warn',\n      'ai-code-snifftest/enforce-domain-terms': 'warn',${archRulesConfig}\n    }\n  },\n  {\n    files: ['**/*.test.js', '**/*.spec.js'],\n    rules: { 'max-lines-per-function': 'off', 'complexity': 'off' }\n  }\n];\n`;
+  fs.writeFileSync(file, content);
+  console.log(`Wrote ${file}`);
+}
+
+module.exports = { writeEslintConfig };
